pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "qr-builder"
version = "0.1.0"
description = "Generate and embed QR codes into images via Python, CLI, and FastAPI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "AIQSO", email = "dev@aiqso.io" }
]
keywords = ["qr", "qrcode", "image", "fastapi", "cli", "branding", "marketing"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Intended Audience :: Developers",
  "Topic :: Multimedia :: Graphics",
  "Framework :: FastAPI",
]

dependencies = [
  "qrcode[pil]>=7.4.2",
  "Pillow>=10.0.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0",
  "httpx>=0.27.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]

[project.scripts]
qr-builder = "qr_builder.cli:main"
qr-builder-api = "qr_builder.api:run"

[project.urls]
Homepage = "https://github.com/aiqso/qr-builder"
Source = "https://github.com/aiqso/qr-builder"
Issues = "https://github.com/aiqso/qr-builder/issues"
Documentation = "https://github.com/aiqso/qr-builder#readme"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.coverage.run]
source = ["qr_builder"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
]
